Adding text input and fetching user input

+++ lib/transaction.dart
@@ -3,7 +3,7 @@ import 'package:flutter/cupertino.dart';
 class Transaction {
   final String id;
   final String title;
-  final int amount;
+  final double amount;
   final DateTime date;
 
   Transaction({
